# backend/confeitaria-spring/openapi.yaml
openapi: 3.0.3
info:
  title: Confeitaria Adriana Novais — API
  description: |
    API REST da confeitaria. Autenticação via **JWT** (Bearer token ou header `x-access-token`).  
    Rotas públicas: `POST /admin/login`, `GET /categorias/**`, `GET /produtos/**`, `POST /cliente`, `POST /pedido`.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local
...
security:
  - bearerAuth: []
  - xAccessToken: []

# ──────────────────────────────────────────────────────────────────────────────
paths:
...
  /produtos:
    post:
      tags: [Produtos]
      summary: Cadastrar produto (com upload de imagem)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProdutoFormData'
      responses:
        '200':
          description: Produto inserido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NovoId'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags: [Produtos]
      summary: Listar produtos
      security: []
      parameters:
        - name: ativo
          in: query
          schema:
            type: boolean
        - name: destaque
          in: query
          schema:
            type: boolean
        - name: categoria_id
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Produto'
        '400':
          $ref: '#/components/responses/BadRequest'
  /produtos/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'

    get:
      tags: [Produtos]
      summary: Buscar produto por ID
      security: []
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'

    put:
      tags: [Produtos]
      summary: Editar produto (com upload de imagem)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/ProdutoFormData'
                - type: object
                  properties:
                    imagem_url:
                      type: string
                      description: URL da imagem existente (mantém se não enviar nova imagem)
      responses:
        '200':
          $ref: '#/components/responses/Mensagem'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'

    delete:
      tags: [Produtos]
      summary: Deletar produto
      responses:
        '200':
          $ref: '#/components/responses/Mensagem'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ── CLIENTES ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido via `POST /admin/login`
    xAccessToken:
      type: apiKey
      in: header
      name: x-access-token
      description: Alternativa ao Bearer — enviar o JWT diretamente no header `x-access-token`
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    Mensagem:
      description: Operação realizada com sucesso
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Mensagem'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Mensagem'
    BadRequest:
      description: Erro de validação ou negócio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
  schemas:
...
    Mensagem:
      type: object
      properties:
        mensagem:
          type: string
    Erro:
      type: object
      properties:
        erro:
          type: string
    NovoId:
      type: object
      properties:
        mensagem:
          type: string
        novo_id:
          type: integer
          format: int64

    # ── Enums ─────────────────────────────────────────────────────────────
...
    CategoriaInput:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
          maxLength: 100
        descricao:
          type: string
          maxLength: 255
        ativo:
          type: boolean
          default: true
    Categoria:
      allOf:
        - $ref: '#/components/schemas/CategoriaInput'
        - type: object
          properties:
            id:
              type: integer
              format: int64
            cadastro:
              type: string
              format: date-time

    # ── Produto ───────────────────────────────────────────────────────────
    ProdutoFormData:
      type: object
      required: [categoria_id, nome, preco, estoque]
      properties:
        categoria_id:
          type: integer
          format: int64
        nome:
          type: string
          maxLength: 150
        descricao:
          type: string
          maxLength: 255
        preco:
          type: number
          format: decimal
        estoque:
          type: integer
          minimum: 0
        destaque:
          type: boolean
          default: false
        ativo:
          type: boolean
          default: true
        unidade_tipo:
          type: string
          maxLength: 20
          example: unidade
        unidade_quantidade:
          type: integer
        imagem:
          type: string
          format: binary
          description: Arquivo de imagem (max 5 MB). A URL pública ficará em `/images/{nome-arquivo}`
    Produto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        categoria:
          $ref: '#/components/schemas/Categoria'
        categoria_id:
          type: integer
          format: int64
        nome:
          type: string
        descricao:
          type: string
        preco:
          type: number
          format: decimal
        estoque:
          type: integer
        imagem_url:
          type: string
          description: Nome do arquivo — acesse via `/images/{imagem_url}`
        destaque:
          type: boolean
        ativo:
          type: boolean
        unidade_tipo:
          type: string
        unidade_quantidade:
          type: integer
        cadastro:
          type: string
          format: date-time

    # ── Cliente ───────────────────────────────────────────────────────────
...